Revert "Feat: navigation bar 구현"
This reverts commit f06896edd50d0632c820df5c8c785065a7b7748f.
