--- components/layout/Layout.tsx
@@ -1,15 +1,0 @@
-import MainNavigation from '@components/layout/MainNavigation';
-
-type MemoonjangLayoutProps = {
-  children: React.ReactNode;
-};
-const Layout = (props: MemoonjangLayoutProps) => {
-  return (
-    <>
-      <main>{props.children}</main>
-      <MainNavigation />
-    </>
-  );
-};
-
-export default Layout;
