--- components/layout/Seo.tsx
@@ -1,11 +1,0 @@
-import Head from 'next/head';
-
-const Seo = ({ title }: { title: string }) => {
-  return (
-    <Head>
-      <title>Me-Moonjang : {title}</title>
-    </Head>
-  );
-};
-
-export default Seo;
