+++ pages/index.tsx
@@ -1,11 +1,10 @@
-import type { GetServerSideProps, NextPage } from 'next';
-import { getSession, useSession } from 'next-auth/react';
-
+import type { NextPage } from 'next';
+import { getSession, signOut, useSession } from 'next-auth/react';
+import Head from 'next/head';
 import Link from 'next/link';
+import axios from 'axios';
 
 import styles from '../styles/Home.module.css';
-import Layout from '@components/layout/Layout';
-import Seo from '@components/layout/Seo';
 
 const Home: NextPage = () => {
   const { data: session } = useSession();
@@ -13,9 +12,13 @@ const Home: NextPage = () => {
 
   return (
     <div className={styles.container}>
-      <Seo title="Home" />
+      <Head>
+        <title>Me-Moonjang</title>
+        <meta name="description" content="me-moonjang" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
 
-      {session ? mainPage() : gate()}
+      {session ? mainPage({ session }) : gate()}
     </div>
   );
 };
@@ -34,17 +37,45 @@ const gate = () => {
   );
 };
 
-const mainPage = () => {
+const mainPage = ({ session }: any) => {
+  const { id } = session.user;
+
+  const handleLogout = async () => {
+    try {
+      const res = await axios.post(`api/auth/logout`, {
+        id,
+      });
+
+      if (res.status === 200) {
+        signOut();
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
   return (
-    <Layout>
-      <main className="container mx-auto text-center mt-20 p-5">
-        <h3 className="text-4xl font-bold">Me-Moonjang</h3>
-      </main>
-    </Layout>
+    <main className="container mx-auto text-center mt-20">
+      <h3 className="text-4xl font-bold">Me-Moonjang</h3>
+
+      <div className="details">
+        <h5>{session.user.name}</h5>
+        <h5>{session.user.email}</h5>
+      </div>
+
+      <div className="flex justify-center">
+        <button
+          onClick={handleLogout}
+          className="mt-5 px-10 py-1 rounded-sm bg-indigo-500"
+        >
+          로그아웃
+        </button>
+      </div>
+    </main>
   );
 };
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
+export async function getServerSideProps(context: any) {
   const session = await getSession(context);
 
   return {
@@ -52,6 +83,6 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
       session,
     },
   };
-};
+}
 
 export default Home;
